refactor(ops): extract Table/Filter/Project/GroupBy ops
New _basic_ops.py (~430 lines) holds the four foundation Op classes:

- SemanticTableOp: leaf relation that carries dimension/measure
  metadata. Uses lazy imports for _make_schema,
  _mutate_dimensions_with_dependencies, and the module logger from
  _core.
- SemanticFilterOp: predicate filtering with derived-dimension
  enrichment. Lazy-imports _Resolver/_get_merged_fields/_resolve_expr/
  _to_untagged/_unwrap and SemanticAggregateOp for its post-agg check.
- SemanticProjectOp + helpers _classify_fields,
  _process_nested_access_marker, _evaluate_measures_with_unnesting,
  _build_select_or_aggregate.
- SemanticGroupByOp: thin pass-through.

All call sites that used to live in _core re-import from _basic_ops
so external `from boring_semantic_layer.ops import X` keeps working.

_core.py is now 2,908 lines (down from 5,330 at start of branch). 930
tests pass.
